+++ app/system/modules/user/src/Controller/UserController.php
@@ -107,7 +107,7 @@ class UserController
     }
 
     /**
-     * @Access("system: manage settings")
+     * @Access("system: access settings")
      */
     public function settingsAction()
     {